+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+## 0.1.2
+* Added support for systems that return or have where clauses
 
 ## 0.1.1 - 2025-06-06
-* Fixed generic assets causing an error, so now `#get_resource<Assets<Mesh>` etc should work.
+* Fixed generic assets causing an error, so now `#get_resource<Assets<Mesh>>` etc should work.